% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-list-su-files.R
\name{rwa_list_su_files}
\alias{rwa_list_su_files}
\title{List Data files in a S3 Bucket}
\usage{
rwa_list_su_files(
  bucket_name,
  data_type = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  file_name = NULL,
  file_extension = NULL,
  max_request_size = 1000,
  silent = FALSE,
  aws_access_key_id = Sys.getenv("AWS_ACCESS_KEY_ID"),
  aws_secret_access_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  region = Sys.getenv("AWS_REGION", "ca-central-1")
)
}
\arguments{
\item{bucket_name}{A string of the AWS S3 bucket name.}

\item{data_type}{A string (by default \code{NULL}) for which data type to return.
Options currently include punch-data, tracks, logger, image and pdf.}

\item{year}{A whole number (by default \code{NULL}) indicating which year is
returned. Format YYYY.}

\item{month}{A whole number (by default \code{NULL}) indicating which month is
returned. Format MM.}

\item{day}{A whole number (by default \code{NULL}) indicating which day is
returned. Format DD.}

\item{file_name}{A string (by default \code{NULL}) containing the name of the file
to be returned. Do not include extension type.}

\item{file_extension}{A string (by default \code{NULL}) for which file extension
to return. Do not include period in front of extension.}

\item{max_request_size}{A whole number (by default \code{1000}) indicating the
maximum number of files to be returned.}

\item{silent}{A flag (by default \code{FALSE}) to toggle messages about number of
files returned on and off. Set to \code{TRUE} to silence messages.}

\item{aws_access_key_id}{A string of your AWS user access key id from AWS.
The default is the system environment variable of \code{AWS_ACCESS_KEY_ID}.}

\item{aws_secret_access_key}{A string of your AWS user secret access key from
AWS. The default is the system environment variable of
\code{AWS_SECRET_ACCESS_KEY}.}

\item{region}{A string of the AWS region you are working within. The default
is the system environment variable of \code{AWS_REGION}}
}
\value{
Vector of object names that match the filter criteria
}
\description{
Wrapper function to \code{rwa_list_files} built to make it easy for accessing
uploaded files from Shiny apps.
}
\examples{
\dontrun{
# Getting all the objects related to punch data
rwa_list_su_files(bucket_name = "my_bucket")

# Get only punch data from 2021
rwa_list_su_files(bucket_name = "my_bucket", year = 2021)

# Get only the csv files
rwa_list_su_files(bucket_name = "my_bucket", extension = "csv")

# Combine to only get excel files submitted in 2021
rwa_list_su_files(bucket_name = "my_bucket", extension = "xlsx", year = 2021)
}
}
