% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-write-files.R
\name{rwa_write_files}
\alias{rwa_write_files}
\title{Save Files to Local Drive}
\usage{
rwa_write_files(file_names, directory, bucket_name)
}
\arguments{
\item{file_names}{A vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}

\item{bucket_name}{A string of the name of the AWS s3 bucket you are trying to access}
}
\description{
Takes a vector of file names and saves them into that directory.
}
