% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for readwriteaws functions}
\arguments{
\item{bucket_name}{A string of the name of the AWS s3 bucket you are trying to access}

\item{year}{Regex filter param}

\item{module_name}{Regex filter param}

\item{file_type}{Regex filter param}

\item{date}{Regex filter param format YYYY-MM-DD}

\item{file_extension}{Regex filter param (think this is better then file_type)}

\item{file_name}{Regex filter param}

\item{file_names}{A vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}
}
\description{
Descriptions of the parameters for readwriteaws functions
}
\keyword{internal}
