% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for readwriteaws functions}
\arguments{
\item{bucket_name}{A character string of the name of the AWS s3 bucket you are trying to access}

\item{pattern}{A character string containing a regular expression to be matched in the given character vector}

\item{max_request_size}{An integer indicating the maximum request size sent to AWS API}

\item{silent}{A logical flag to turn on and off messages about number of files returned}

\item{year}{A numeric value for what years you want returned. Format YYYY}

\item{month}{A numeric value for what month you want returned. Format MM}

\item{day}{A numeric value for what day you want returned. Format DD}

\item{data_type}{A character string for which data you want to return. Options currently include punch-data, tracks, logger, image and pdf.}

\item{file_extension}{A character string for which file extension you want without the period}

\item{file_name}{A character string for which file names you want}

\item{file_list}{A vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}
}
\description{
Descriptions of the parameters for readwriteaws functions
}
\keyword{internal}
