% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameters for readwriteaws functions}
\arguments{
\item{bucket_name}{A character string of the name of the AWS s3 bucket you are trying to access}

\item{pattern}{A character string containing a regular expression to be matched in the given character vector}

\item{max_request_size}{An integer indicating the maximum request size sent to AWS API}

\item{silent}{A flag to turn on and off messages about bucket object numbers}

\item{year}{Regex filter param}

\item{month}{Regex filter param - format MM}

\item{day}{Regex filter param - format DD}

\item{data_type}{Regex filter param}

\item{file_extension}{Regex filter param}

\item{file_name}{Regex filter param}

\item{file_names}{A vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}
}
\description{
Descriptions of the parameters for readwriteaws functions
}
\keyword{internal}
