% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-download-files.R
\name{rwa_download_files}
\alias{rwa_download_files}
\title{Download Files from AWS to Local Drive}
\usage{
rwa_download_files(file_list, directory, bucket_name, progress = TRUE)
}
\arguments{
\item{file_list}{A vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}

\item{bucket_name}{A character string of the name of the AWS s3 bucket you are trying to access}

\item{progress}{A logical flag to turn on and off the progress bar}
}
\description{
Copies files from your AWS s3 bucket to your local drive
}
