% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-get-data.R
\name{rwa_get_data}
\alias{rwa_get_data}
\title{List Punch Data Files from s3 Bucket}
\usage{
rwa_get_data(
  bucket_name,
  data_type = NULL,
  year = NULL,
  month = NULL,
  day = NULL,
  file_name = NULL,
  file_extension = NULL,
  max_request_size = 1000,
  silent = FALSE
)
}
\arguments{
\item{bucket_name}{A character string of the name of the AWS s3 bucket you are trying to access}

\item{data_type}{Regex filter param}

\item{year}{Regex filter param}

\item{month}{Regex filter param - format MM}

\item{day}{Regex filter param - format DD}

\item{file_name}{Regex filter param}

\item{file_extension}{Regex filter param}

\item{max_request_size}{An integer indicating the maximum request size sent to AWS API}

\item{silent}{A flag to turn on and off messages about bucket object numbers}
}
\value{
Vector of object names that match the filter criteria
}
\description{
Lists only punch data files
}
\examples{
\dontrun{
# Getting all the objects related to punch data
rwa_get_data(bucket_name = "my_bucket")

# Get only punch data from 2021
rwa_get_data(bucket_name = "my_bucket", year = 2021)

# Get only the csv files
rwa_get_data(bucket_name = "my_bucket", extension = "csv")

# Combine to only get excel files submitted in 2021
rwa_get_data(bucket_name = "my_bucket", extension = "xlsx", year = 2021)
}
}
