% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-get-punch.R
\name{rwa_get_punch}
\alias{rwa_get_punch}
\title{List Files in S3 Bucket}
\usage{
rwa_get_punch(year = NULL, extension = NULL)
}
\arguments{
\item{year}{For narrowing scope of search if used in read-files it will need to build itself up}
}
\description{
Lists all files from s3 bucket
}
