% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-list-files.R
\name{rwa_list_files}
\alias{rwa_list_files}
\title{List Files in S3 Bucket}
\usage{
rwa_list_files(bucket_name, max_request_size = 1000, pattern = ".*")
}
\arguments{
\item{bucket_name}{A string of the name of the AWS s3 bucket you are trying to access}
}
\description{
Lists all files from s3 bucket
}
