% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-read-files.R
\name{rwa_read_files}
\alias{rwa_read_files}
\title{List Files in S3 Bucket}
\usage{
rwa_read_files(bucket_name, year = "", module_name = "", file_type = "")
}
\arguments{
\item{bucket_name}{A string of the name of the AWS s3 bucket you are trying to access}

\item{year}{For narrowing scope of search if used in read-files it will need to build itself up}

\item{module_name}{For scope like tracks, pdf, etc}

\item{file_type}{Regex later for grabbing what is wanted}
}
\description{
Lists all files from s3 bucket
}
