% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwa-upload-files.R
\name{rwa_upload_files}
\alias{rwa_upload_files}
\title{Upload Files from your Local Drive to AWS}
\usage{
rwa_upload_files(
  file_list,
  directory = "",
  bucket_name,
  bucket_path = "",
  aws_access_key_id = Sys.getenv("AWS_ACCESS_KEY_ID"),
  aws_secret_access_key = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  region = Sys.getenv("AWS_REGION", "ca-central-1")
)
}
\arguments{
\item{file_list}{A character vector containing a list of files from the object names in AWS}

\item{directory}{A string that names the folder you want to save the objects into.}

\item{bucket_name}{A string of the name of the AWS s3 bucket you are trying to access}

\item{bucket_path}{A character string for adding a different folder to where the files are saved in your bucket. Default it "" which means it saves in the directory given by file_list and directory.}

\item{aws_access_key_id}{Your AWS user access key id that. The default is to use your system environment variable of \code{AWS_ACCESS_KEY_ID}}

\item{aws_secret_access_key}{Your AWS user secret access key that. The default is to use your system environment variable of \code{AWS_SECRET_ACCESS_KEY}}

\item{region}{The region you want to create your resource in. The default is to use your system environment variable of \code{AWS_REGION}}
}
\description{
Upload files from your local drive to an AWS s3 Bucket
}
